requirements.txt
pyosmeta
pandas
altair
-e ./pyosmetrics_pkg
jupyter
pyarrow
itables # interactive tables!
tqdm
nox
pygithub # for github API
